application/requirements.txt
APScheduler==3.11.0
cryptography==46.0.1
expiringdict==1.2.2
fastapi==0.117.1
Flask==3.1.2
flask_sqlalchemy==3.1.1
keyring==25.6.0
pydantic==2.11.9
python-telegram-bot==22.4
pythonping==1.1.4
pytz==2025.2
Requests==2.32.5
scapy==2.6.1
SQLAlchemy==2.0.43
